[CHORE] 코드 리뷰 반영 - extension 파일 분리
