+++ Hatnya/Screens/CreateStudy/UIComponent/DayButtonStackView.swift
@@ -43,4 +43,5 @@ class DayButtonStackView: UIStackView {
             addArrangedSubview(btn)
         }
     }
+    
 }
