+++ Hatnya/Screens/CreateStudy/UIComponent/SelectCycleDaysView.swift
@@ -8,17 +8,16 @@
 import UIKit
 
 class SelectCycleDaysView: UIView {
-    
     let cyclePickerView = CyclePickerView()
     let dayButtonStackView = DayButtonStackView()
     
-    let selectedCycleDaysLabel: UILabel = {
-        let label = UILabel()
-        label.text = "숙제가 매주 월요일 및 목요일에 반복됩니다."
-        label.font = UIFont.systemFont(ofSize: 14)
-        label.textColor = .systemGray2
-        return label
-    }()
+//    let selectedCycleDaysLabel: UILabel = {
+//        let label = UILabel()
+//        label.text = "숙제가 매주 월요일 및 목요일에 반복됩니다."
+//        label.font = UIFont.systemFont(ofSize: 14)
+//        label.textColor = .systemGray2
+//        return label
+//    }()
     
     override init(frame: CGRect) {
         super.init(frame: frame)
@@ -46,17 +45,60 @@ class SelectCycleDaysView: UIView {
         NSLayoutConstraint.activate([
             dayButtonStackView.topAnchor.constraint(equalTo: cyclePickerView.bottomAnchor, constant: 20),
             dayButtonStackView.leadingAnchor.constraint(equalTo: cyclePickerView.leadingAnchor),
-            dayButtonStackView.centerXAnchor.constraint(equalTo: cyclePickerView.centerXAnchor)
+            dayButtonStackView.centerXAnchor.constraint(equalTo: cyclePickerView.centerXAnchor),
+            dayButtonStackView.bottomAnchor.constraint(equalTo: self.bottomAnchor)
         ])
         
-        addSubview(selectedCycleDaysLabel)
-        selectedCycleDaysLabel.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            selectedCycleDaysLabel.topAnchor.constraint(equalTo: dayButtonStackView.bottomAnchor, constant: 20),
-            selectedCycleDaysLabel.leadingAnchor.constraint(equalTo: dayButtonStackView.leadingAnchor),
-            selectedCycleDaysLabel.centerXAnchor.constraint(equalTo: dayButtonStackView.centerXAnchor),
-            selectedCycleDaysLabel.bottomAnchor.constraint(equalTo: self.bottomAnchor)
-        ])
+//        addSubview(selectedCycleDaysLabel)
+//        selectedCycleDaysLabel.translatesAutoresizingMaskIntoConstraints = false
+//        NSLayoutConstraint.activate([
+//            selectedCycleDaysLabel.topAnchor.constraint(equalTo: dayButtonStackView.bottomAnchor, constant: 20),
+//            selectedCycleDaysLabel.leadingAnchor.constraint(equalTo: dayButtonStackView.leadingAnchor),
+//            selectedCycleDaysLabel.centerXAnchor.constraint(equalTo: dayButtonStackView.centerXAnchor),
+//            selectedCycleDaysLabel.bottomAnchor.constraint(equalTo: self.bottomAnchor)
+//        ])
     }
     
 }
+
+// 2차 스프린트 예정 - 선택한 주기 보여주는 라벨
+
+//extension SelectCycleDaysView: UIGestureRecognizerDelegate {
+//    func gestureRecognizer(_ gestureRecognizer: UIGestureRecognizer, shouldReceive touch: UITouch) -> Bool {
+//        var completeSentence = ""
+//        let selectedCycle = cyclePickerView.selectedCycleLabel.text ?? ""
+//
+//        var days: [String] = []
+//
+//        if dayButtonStackView.sundayButton.isSelected {
+//            days.append("일")
+//        }
+//        if dayButtonStackView.mondayButton.isSelected {
+//            days.append("월")
+//        }
+//        if dayButtonStackView.tuesdayButton.isSelected {
+//            days.append("화")
+//        }
+//        if dayButtonStackView.wednesdayButton.isSelected {
+//            days.append("수")
+//        }
+//        if dayButtonStackView.thursdayButton.isSelected {
+//            days.append("목")
+//        }
+//        if dayButtonStackView.fridayButton.isSelected {
+//            days.append("금")
+//        }
+//        if dayButtonStackView.saturdayButton.isSelected {
+//            days.append("토")
+//        }
+//
+//        completeSentence = "숙제가 " + selectedCycle + "마다 "
+//        days.forEach { day in
+//            completeSentence = completeSentence + day + ", "
+//        }
+//        completeSentence = completeSentence + "에 반복됩니다."
+//        selectedCycleDaysLabel.text = completeSentence
+//        return true
+//    }
+//
+//}
